FreeTrialAd to TS, Story refactored and updated.
