+++ .nvmrc
@@ -1,1 +1,1 @@
-14.16.0
+16.10
