+++ src/index.js
@@ -33,7 +33,7 @@ import EmailForm from './components/Forms/EmailForm';
 import FeatureCard from './components/Cards/FeatureCard/FeatureCard';
 import FilterButton from './components/FilterButton/FilterButton';
 import FinePrint from './components/Articles/FinePrint';
-import FreeTrialAd from './components/Ads/ShowcaseAds/FreeTrialAd';
+import FreeTrialAd from './components/Ads/ShowcaseAds/FreeTrialAd/FreeTrialAd';
 import HeroAd from './components/Ads/HeroAd/HeroAd';
 import HeroCard from './components/Cards/HeroCard';
 import HomepageTagline from './components/HomepageTagline';
